backend/requirements.txt
Django~=4.1.1
psycopg2-binary~=2.9.3
django-debug-toolbar~=3.6.0
graphene-django==3.0.0
